MAYA-123876 preserve session layer on save
When a new stage with an anonymous root layer is saved for teh first time, the session layer was lost.

The reason was that when a new stage is saved, its root layer change ID. When the proxy shape was recomputed, it created tried to find the root layer with teh new ID in the stage cache. Since the just-saved stage was previously using an anonymous root layer, the cache does not find the just-saved stage. A new stage got created, which meant it got a brand new empty session layer.

This recomputation happened mid-save.

The fix is to avoid recomputing the proxy stage mid-save and to pre-fill the stage cache with the newly-saved root with its new ID and with the pre-existing session layer. This requires keeping more information when initially gathering the layers to be saved to avoid touching the proxy shape again during the save.

- Add BatchSaveInfo to the layer manager to cache all info necessary during saving.
- Pass these info instead of only the DAG path to callbacks and various internal functions.
- Have the LayerDatabase fill this info when gathering layers to save.
- Avoid calling getProxiesToSave() mid-save.
- Add hasDirtyLayer() to support the previous point.
- Refactor clearing proxies to save into its own funciton to avoid code duplication.
- Add GetAllCaches() to UsdMayaStageCache to allow access to all caches without knowing the details of how caches are segregated.
- When saving a layer, prefill all stage caches that were holding the stages that were using the original layer to now cache a stage that uses the new layer and the session layer of the existing stage.
- This will allow the proxy shape to find these stage with teh correct session layer when recomputed later.
- Update the batch save layer UI to the new API with BatchSaveInfo.