MAYA-123876 save the session layer when saving to external USD files
The layer manager would not save the session layer of the stage in the Maya scene file when the layers were being saved to external USD files.

There were two problems: the session layer name was not being set on the stage proxy node and the layer manager was not saved to the maya scene.

The layer manager node would be lost when saving layer to separate USD files. (Actually, this one but one possible case where the layer manager would be lost.) The reason was that the layer manager node would be deleted mid-save because some other code happened to access the stage while the stage was dirty, which would cause the DG to re-compute the stage node, which would access layers in the layer manager, which would wipe the layer manager node. And all this was dependent on the order some save callbacks were being called.

To avoid this, we now no longer wipe the layer manager node when in the middle of a save operation.

Also, improved the proxy shape test to cleanup after themselves, deleteing their temp files.