Add API writers
Some Maya nodes may not correspond directly to Usd prims. One example of
this is Renderman-For-Maya's PxrMeshLight, which should be exported as
an applied MeshLightAPI on a Mesh prim.

To support this, this change adds an apiWriterRegistry that allows us to
register apiWriters that get run for each exported prim. We also want to
prevent the export from creating a node for these Maya nodes, so we now
allow a type to be registered as a type that should not result in a Usd
prim.