MAYA-125101 Correct handling of display list vs variants
When undoing the deletion of an ancestor of a USD prim with variant which had some of its children edited as Maya, we need to validate that the correct set of variant is active before showing the edited Maya data.

Also, when going down the items of the trie that contains the UFE path component, we need to properly handle the points where we cross run-times (currenly there is only Maya and USD). When crossing runtimes, we must create a new segment.