MAYA-125101 void requesting invalid object
When in the middle of deleting a proxy shape in the outliner, the outliner updates itself and checks if nodes are gateway nodes and that access the stage map where the just-deleted proxy shape is still in the map and triggered an assertion. Avoid trying to access the object associated with stale entries.