Properly track usage of shared render items
In #1339 code was added to share the selection render item between
reprs. The code did not track usage, so as soon as one of the reprs is
removed, the render item was removed from the subscene as well, leaving
dangling pointers in the remaining reprs.

In #1597 code was added that seem to indicate there was suspicion about
render item issues, but sharing was not identified as the cause.

This PR adds use count tracking to the render item so they only get
removed from the subscene once all reprs have stopped using the shared
render item.