MAYA-125867 fix edit-as-Maya on fresh Maya session
When a USD object is edited as Maya and the Maya scene is saved, if Maya is restarted and teh scene reloaded, the edited USD object cannot be manipulated again. That was due to internal non-saved state not being set properly when restarting Maya.

Avoid the problem by not using non-persistent state.