MAYA-122481 fix rendering crash with MtoH
When rendering, with some options, the renderer adds atemporary light to the scene, then removes it as soon as the render ends. The Maya scene delegate receives scene addition notifications, but in this case by the time the consequences if the notifications are processed, the light has already been deleted and the delegate access freed memory.

Make the scene delegate also listen to scene item removal notifications to remove items that have been deleted before they can be processed.