MAYA-125036 correctly disable merge to USD
When two variants contain prim with the same name, editing one prim in one variant then switching to the other variant would still show the "Merge Maya Edits to USD" menu item enabled due to the confusion of still having a prim with the same name.

Augment the code to properly detect that the prim has no pull information and disable the merge-to-USD menu item.