MAYA-126483 Fix a bug where materials got created in the wrong scope.
When using "Assign New Material", a new material gets created in the default materials scope. This
scope is usually called "mtl". In the previous implementation, there was a bug where materials got
created in *any* scope that starts with "mtl". If, for example, there was a scope called "mtlFoo",
and no scope called "mtl", the new material got created in "mtlFoo". However, the expected behaviour
in this example is to create a scope called "mtl" and to create the new material in there.

This is the expected behaviour in a more general sense: If the default materials scope name (usually
"mtl") is available, create the new material in a scope with this name.
If the default materials scope name is not available i.e., there is already something that is not
of type "Scope" that is using that name (e.g. an Xform called "mtl"), append a number suffix to the
default materials scope name (e.g. "mtl1") and try to use that. If this name is not available
either, increment the number suffix until an available name is found.