MAYA-122917 interaction between deactivation and cancel edit
When a Maya reference is edited, its USD prim is deactivated to hide it in the outliner. If its parent prim is then deactivated, the edit is orphaned. At that point the Maya Ref prim and its parent are both deactivated. If the edit is canceled while the parent is deactivated, the code could not activate the Maya Ref USD prim because children of deactivated prims cannot be modified. We needed to fix this.

The fix is to add code to temporarily reactivate ancestors prims when trying to activate a prim. The design actuslly provides a class to temporarily activate all ancestors of a prim, for any possible edits.

- Add PrimActivation class to temporarily make a prim accessible even when its ancestors are deactivated.
- It works by activating all inactive ancestors by authoring opinions in the session layer.
- Use this class in the addExcludedFromRendering and removeExcludedFromRendering functions.
- Add a flag in the orphaned nodes manager to avoid responding to notifications caused by its own responses to notifications.
- This avoids infinite recursion when the work done to orphan an edited node caused notifications to be sent.
- In our case, the temporary reactivation of prims would cause an infinite flip-flop or orphaning and de-orphaning.