MAYA-122927 re-use unit test
A unit test already checked the case where an edited prim was under a dectivated prim, but had not checked for the proper result due to the known bug. Simply fix the test for the fix.

Also, fix that the auto-edit flag was not being properly updated due to the same ancestor-is-deactivated problem.