MAYA-127045 Add reparenting rules for UsdShadeShader, UsdShadeNodeGraph and UsdShadeMaterial.
Users are able to reparent USD items in the Outliner. However, in some cases, reparenting does not
make sense and can lead to problematic situations in LookdevX. The following rules are established,
when a user tries to reparent items in the Outliner:

- UsdShadeShader can only have UsdShadeNodeGraph/UsdShadeMaterial as parent.
- UsdShadeNodeGraph can only have a UsdShadeNodeGraph/UsdShadeMaterial as parent.
- UsdShadeMaterial can not have UsdShadeShader/UsdShadeNodeGraph/UsdShadeMaterial as parent.

If one of the rules is broken, reparenting is prevented.