MAYA-127095 remove usage of macros leading to fatal exits
We should strive to reduce the number of scenarios under which Maya crashes.
Instead, when possible, we should report an error to user and abort the
operation. MayaUSD has extensive undo/redo capabilities now and code is able to
undo partially done operations. Given this, errors should be reported and
propagated back.

The macros TF_AXIOM and TF_FATAL_ERROR cause crashes when their conditions are
not met, even in release builds. Remove TF_AXIOM and TF_FATAL_ERROR, replacing
them with:

- TF_WARN when the situation was not an error but an unusual situation.
- TF_VERIFY when it still warrant an error message to the user, for example during error handling.
- Simple if/else, with an error return value when not a hard error.
- Simple if when the condition was trivial, like an empty node when iterating over USD nodes.
- In a few cases, TF_DEV_AXIOM when the condition was really strictly enforced by a class constructor.
- Entirely removed, when the code was already dealing with the error.
- C++ exception when the error needs to propagate to caller, in particular in constructors.