MAYA-125816 fix USD export curve options
When the option is enabled, it should get its default value if not explicitly set.

The previous fix aim was to avoid disabling the UI when the option is absent, but it also prevented setting the option to its default value. The fix is to fill the default value when the UI is enabled.