MAYA-127478 fix bulk save of root layer filename
When doing a bulk-save of layers when saving a Maya scene for the first time, the multiple-parts of teh saving process could end-up using stale data. That is because there is a first batch-save delegate (customizable callback) that can do a partial save, then a fallback save is done for any layer that did not get saved.

When a stage has just been created, it contains an anonymous root layer which gets replaced by a non-anonymous by the batch save delegate. When the root layer is swapped, the stage proxy shape is updated by setting its stage name attribute.  Unfortunately, the USD stage object was being cached and the attribute chage was thus not use immediately. Thus the 2nd part fallback-save thought that the root layer was still anonymous.

The fix is to refresh the cached USD stages by pulling again on the proxy stage node so that it updates itself when its attributes have been changed.