MAYA-124737 improve duplicate command
The concept of layer from within the USD API is confusing relative to the high-level concept of layers. When talking about layers at a high-level, what is meant are the root and session layers and their sub-layers. OTOH, in the API, the layers are also used for references.  A reference is kept in a layer object (SdfLayer), but is not part of the stage layer stack!

In the case of duplication, we use SdfCopySpec to do the heavy lifting, and that takes a source path and source layer. But these are low-level layers, not high-level layers. n particular, if the source prim is in a reference, we must use the reference layer as the layer and the path must be relative to reference, not the stage root!

So, change the duplicate code to handle references correctly.

- Add a helper function to retrieve the defining layer and path of a prim.
- The defining path may not be the prim path if the prim is in a reference.
- In that case, the path is relative to the reference target, in the referenced layer.
- Also, refactor the defPrimSpecLayer function to make each check alone and separate to clarify and help when debugging.
- Remove the restriction on duplicate: a new prim is about to be created, it does not care about the structure of the source prim, so it should not be restricted.
- Use the new helper to find the corretc source layer and path.
- Make sure to create missing parent 'overs'.
- Make the error message clearer by making it about USD prim, not low-level prim spec that the user might not know about.
- Make duplicate work with multi-layers prims by iterating over all layers with opinions and merge all opinions into the destination.
- Add helper function to retrieve all layer and path that have opinions.
- Add a unit test for the duplication multi-layers prim.
- Add a unit test for the duplication of a prim in a reference.