MAYA-124922 use Load/Unload for payload
Unfortunately, USD stage's SetLoadRules function does not optimize the change notifications it sends out. Its Load and Unload functions do optimize the notifications they send out. Normally, this is an implementation details inside USD that we should not have to rely on. Pragmatically though, using the functions which send out the most narrow notification is better. So even though calling SetLoadRules is simpler and more precise, it is less efficient.