MAYA-126128 sync session layer for various commands
Keep opinions in the session layer in sync when a prim is renamed while
targeting another layer. Do the same for insert child, group delete and reorder
commands.

Provide helper functions to deal withteh session layer and enhance the various
commands to allow them when the session layer contains opinions, by using these
helper functions.

Allow authoring in session layer: detect that the target is in the session
layer and allow edits there if it is the strongest session opinion.

- Move many layer-related helpers into their own file.
- Refactor code that used the old functions to use the new functions.
- Add applyToAllPrimSpecs and applyToAllLayersWithOpinions helper functions.
- Add applyToSomeLayersWithOpinions helper function to apply a change to a subset of layers.
- Use applyToAllPrimSpecs helper function in the rename command.
- Use the helper functions in the rename, delete, insert child and reorder commands..
- Properly delete all prim spec in session layer when deleting a prim.
- Properly throw an exception when failing.
- This is necessary so that composite commands know a sub-command has failed.
- Add rename unit test that demonstrate that prims can be renamed when they have opinions in the session layer.
- Add unit tests for group (insert child) and reorder.
- Fix the delete command unit test.
- Fix unit tests to use non-session layer when testing restrictions.