MAYA-127817 make stage set by ID reloadable
When the stage is set on the proxy shape via the stage ID attribute, fill the file path attribute automatically so that the stage can be reloaded when teh Maya scene file is reloaded.

The stage ID attribute is not saved and the ID is a run-time value in USD that is not constant from run to run. The only way to be reloadeable is to save the file path to the root layer.

- Fill the file path attribute when computing the stage if the stage is set by stage ID.
- Only do this if the stage is not anonynous.
- Add a unit test that verifies the stage is reloaded when the Maya scene is reloaded.