Fix unit test
Anything that creates a MPlug on the proxy shape will end up causing a
leak that will help the USD stage survive the File->New/Open cycle. This
leak has been there for quite a while, so let's make sure we are not
re-using a tainted prim in the second test.

No need to even create a stage listener. The action of getting a plug on
the stage is sufficient to re-create that leak.