MAYA-128513 fix load rules duplication
The Load rules are not part of the USD data that gets restore by automatic USD undo block, so we need to dulcaute them explicitly. Also, the UFE reparent notification need to be triggered during undo and redo.

The unit tests were tied to some subtle behavior of the implementation. Now that we use "perfect" undo/redo, the order f items is also preserved, so the testGroupCmd assertions had to be fixed. Similarly, the parent command was assuming the parent prim was passed to the edit router. It was also creating the parent prims, which is not necessary. (The problem was fixed before, but the fix was hidden due to the exact timing of the edit routing call.)

Also made the parent unit test for edit routing properly unregister the router even when the test fails. Otherwise, the subsequent test crashes due to the invalid edit router being left behind.