HYDRA-339 - fix the calculation of pickmatrix and use a smaller pick buffer (#409)
HYDRA-339 - fix the calculation of pickmatrix and use a smaller pick buffer

Changes:

Fix the error in calculation of pick matrix
What pick matrix does is applies a scale and translation to the
projection matrix so that a small part of the scene around the pick point gets projected into
the whole viewport. Most of the scene will get clipped away. What doesn't get clipped is a candidate for 'picking'.
The calculation of pick matrix can be deducted by:
1). map the given viewport sub region around pick point into clip space (assuming that the given viewport covers -1…1)
2). compute a matrix which scales and translates this clip space sub region to cover the whole area of -1…1 in X and Y

Use the size of pick region instead of full viewport to render id buffers for picking to improve perf
1).Marquee/single click selection would get improvement as the pick region is usually much smaller than full viewport.
2).Point snapping selection wouldn't get improvement as the pick region is same as full viewport, need to think about other approaches.

Impacts:
Selection &Highlighting under mayaHydra