MAYA-129059 - Add command to create stages.
Add an undoable command that creates a stage with a new layer.

The new command is a C++ implementation of the existing mayaUsd_createStageWithNewLayer.py Python script. The C++ command will:
- enable the creation of stages from within the C++ code.
- enable the creation of stages deep in the Maya hierarchy.
- wrap stage creation into an Ufe::UndoableCommand so that it can be undone by UFE clients.

This commit also adds a Python binding for the stage creation command, which is used in mayaUsd_createStageWithNewLayer.py to replace the current implementation.