Use Hd_VertexAdjacency API directly
Updated to use the vertex adjacency API directly to build the vertex
adjacency table instead of using Storm specific shared builder
computations.

This simplifies the code sightly while avoiding a dependency on API
that is being moved from imaging/hd to imaging/hdSt.

Not using a version guard here since the interface we're switching to
has been available for a long time.