MAYA-128939 remove error message on deleting a stage
The error was printed from a utility function to map from a DAG object to a USD stage. Utility functions should not warn, the caller should warn if it realyl thinks it is an error. Here, we are responding to a notification sent in the middle of DAG object deletion, so it is expected that the object is not valid. This shows that printing an error in this case is incorrect.