MAYA-128503 - Improve roundtripping of GeomSubsets
We have two sources of GeomSubsets in Maya:
 - Shaded faces
 - Explicitly create component tags

In USD land GeomSubsets can be modified in ways that have no Maya
equivalent:
 - We can change the family name
 - We can assign a material

So when importing/exporting USD data, we need to add just enough info
that all Maya and USD sepcificities are accounted for and correctly
handled.

The unit test will be quite informative.